chore: remove `:error_handlers` from required fields
this fixes the following warning:
attempting to determine the presence of embed_many field :error_handlers with validate_required/3 or field_missing?/2 which has no effect. You can pass the :required option to Ecto.Changeset.cast_embed/3 to achieve this.

as we already have `required: true` in the `cast_embed`, this should
be a non-issue

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>